made route for getting currently injured players
