+++ nba-app/server/routes/players.js
@@ -2,8 +2,7 @@ import express from 'express';
 import Team from '../models/Team.js'; // Adjust the import based on your file structure
 import Player from '../models/Player.js'
 import path from 'path';
-
-const scriptPath = path.resolve('../dataPopulation/nbainjuries.py');
+import setInjuredPlayers from "../dataPopulation/injuryUpdate.js";
 
 const router = express.Router();
 
@@ -27,8 +26,32 @@ router.get('/:teamName/players', async (req, res) => {
   }
 });
 
-router.get('/:teamName/injuredPlayers', (req, res) => {
+router.get('/:teamName/injuredPlayers', async (req, res) => {
     const { teamName } = req.params;
+
+    try {
+      // First, update the injury status of players by calling setInjuredPlayers
+      await setInjuredPlayers(teamName);
+  
+      // Find the team by the given team name
+      const team = await Team.findOne({ name: teamName });
+  
+      if (!team) {
+        return res.status(404).json({ message: 'Team not found' });
+      }
+  
+      // Find all injured players from the team
+      const injuredPlayers = await Player.find({
+        _id: { $in: team.players },  // Only search for players in the team
+        isInjured: true              // Only retrieve injured players
+      });
+  
+      // Return the list of injured players
+      res.status(200).json(injuredPlayers);
+    } catch (error) {
+      console.error(`Error fetching injured players: ${error.message}`);
+      res.status(500).json({ message: 'Server error' });
+    }
   
     
   });
